requirements.txt
# Core Flask Dependencies
flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
pillow==11.3.0
numpy==2.2.6
werkzeug==3.1.3

# AI Background Removal - Core
rembg==2.0.67
onnxruntime==1.19.2

# Scientific Computing - Essential only
scipy==1.14.1
opencv-python-headless==4.10.0.84

# Supporting Libraries - Minimal
requests==2.32.4
tqdm==4.67.1
flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
requests==2.32.4
tqdm==4.67.1
werkzeug==3.1.3
numpy
onnxruntime
opencv-python-headless
pillow==11.3.0
rembg
scipy